# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Customer Review Intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: reviews
    description: Review operations
  - name: trends
    description: Trend analysis operations
  - name: reports
    description: Report generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/compare:
    get:
      operationId: compareProducts
      tags: [products]
      summary: Compare products
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comparison matrix
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComparisonMatrix"

  /products/{id}/overview:
    get:
      operationId: getProductOverview
      tags: [products]
      summary: Get product overview metrics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Overview metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductOverview"

  /products/{id}/features:
    get:
      operationId: getProductFeatures
      tags: [products]
      summary: Get per-feature aggregates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feature aggregates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeatureAggregate"

  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [reviews]
      summary: Get paginated product reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: filter
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"

  /products/{id}/trends:
    get:
      operationId: getProductTrends
      tags: [trends]
      summary: Get trend snapshots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: feature
          in: query
          schema:
            type: string
        - name: window
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendSnapshot"

  /products/{id}/trends/emerging:
    get:
      operationId: getEmergingIssues
      tags: [trends]
      summary: Get emerging issues
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Emerging issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergingIssue"

  /products/{id}/trends/anomalies:
    get:
      operationId: getAnomalies
      tags: [trends]
      summary: Get anomaly timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anomaly timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anomaly"

  /reviews/upload:
    post:
      operationId: uploadReviews
      tags: [reviews]
      summary: Upload reviews file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                product_id:
                  type: integer
      responses:
        "200":
          description: Ingestion summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionSummary"

  /reviews/manual:
    post:
      operationId: analyzeReview
      tags: [reviews]
      summary: Analyze a single review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualReviewBody"
      responses:
        "200":
          description: Analyzed review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/stream:
    post:
      operationId: streamReview
      tags: [reviews]
      summary: Stream a single review for analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualReviewBody"
      responses:
        "200":
          description: Analyzed review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reports/{product_id}/pdf:
    get:
      operationId: getReportPdf
      tags: [reports]
      summary: Download PDF report
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF blob
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  /reports/{product_id}/json:
    get:
      operationId: getReportJson
      tags: [reports]
      summary: Get structured JSON report
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Structured report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        total_reviews:
          type: integer
        valid_reviews:
          type: integer
        positive_pct:
          type: number
        negative_pct:
          type: number
        neutral_pct:
          type: number
        languages:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - total_reviews
        - valid_reviews
        - positive_pct
        - negative_pct
        - neutral_pct
        - languages
        - created_at

    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
      required:
        - name
        - category

    ProductOverview:
      type: object
      properties:
        total_reviews:
          type: integer
        valid_reviews:
          type: integer
        spam_count:
          type: integer
        duplicate_count:
          type: integer
        sarcastic_count:
          type: integer
        emerging_issues_count:
          type: integer
        languages_detected:
          type: array
          items:
            type: string
        language_breakdown:
          type: object
          additionalProperties:
            type: integer
        overall_sentiment:
          type: object
          properties:
            positive:
              type: number
            neutral:
              type: number
            negative:
              type: number
            ambiguous:
              type: number
          required:
            - positive
            - neutral
            - negative
            - ambiguous
      required:
        - total_reviews
        - valid_reviews
        - spam_count
        - duplicate_count
        - sarcastic_count
        - emerging_issues_count
        - languages_detected
        - language_breakdown
        - overall_sentiment

    FeatureAggregate:
      type: object
      properties:
        feature:
          type: string
        mention_count:
          type: integer
        positive_pct:
          type: number
        neutral_pct:
          type: number
        negative_pct:
          type: number
        ambiguous_pct:
          type: number
      required:
        - feature
        - mention_count
        - positive_pct
        - neutral_pct
        - negative_pct
        - ambiguous_pct

    Review:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        text:
          type: string
        language:
          type: string
        overall_sentiment:
          type: string
          enum: [positive, negative, neutral, ambiguous]
        is_spam:
          type: boolean
        is_duplicate:
          type: boolean
        is_sarcastic:
          type: boolean
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureSentiment"
        created_at:
          type: string
          format: date-time
        review_date:
          type: string
      required:
        - id
        - product_id
        - text
        - language
        - overall_sentiment
        - is_spam
        - is_duplicate
        - is_sarcastic
        - features
        - created_at

    FeatureSentiment:
      type: object
      properties:
        feature:
          type: string
        sentiment:
          type: string
          enum: [positive, negative, neutral, ambiguous]
        confidence:
          type: number
        evidence:
          type: string
      required:
        - feature
        - sentiment
        - confidence

    PaginatedReviews:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer
      required:
        - items
        - total
        - page
        - size
        - pages

    TrendSnapshot:
      type: object
      properties:
        index:
          type: integer
        date:
          type: string
        feature:
          type: string
        negative_pct:
          type: number
        positive_pct:
          type: number
        mention_count:
          type: integer
      required:
        - index
        - date
        - feature
        - negative_pct
        - positive_pct
        - mention_count

    EmergingIssue:
      type: object
      properties:
        feature:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        from_pct:
          type: number
        to_pct:
          type: number
        is_systemic:
          type: boolean
        review_count:
          type: integer
        description:
          type: string
      required:
        - feature
        - severity
        - from_pct
        - to_pct
        - is_systemic
        - review_count
        - description

    Anomaly:
      type: object
      properties:
        date:
          type: string
        feature:
          type: string
        delta:
          type: number
        description:
          type: string
      required:
        - date
        - feature
        - delta
        - description

    IngestionSummary:
      type: object
      properties:
        ingested:
          type: integer
        duplicates_flagged:
          type: integer
        spam_flagged:
          type: integer
        sarcastic_flagged:
          type: integer
      required:
        - ingested
        - duplicates_flagged
        - spam_flagged
        - sarcastic_flagged

    ManualReviewBody:
      type: object
      properties:
        text:
          type: string
        product_id:
          type: integer
        language:
          type: string
      required:
        - text
        - product_id

    ComparisonMatrix:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        features:
          type: array
          items:
            type: string
        matrix:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: number
      required:
        - products
        - features
        - matrix

    Report:
      type: object
      properties:
        product_id:
          type: integer
        generated_at:
          type: string
          format: date-time
        overview:
          $ref: "#/components/schemas/ProductOverview"
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureAggregate"
        emerging_issues:
          type: array
          items:
            $ref: "#/components/schemas/EmergingIssue"
      required:
        - product_id
        - generated_at
        - overview
        - features
        - emerging_issues