# lib/api-zod/src/generated/types/anomaly.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Customer Review Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Anomaly {
  date: string;
  feature: string;
  delta: number;
  description: string;
}